+++ Cloudy.xcodeproj/xcshareddata/xcschemes/Cloudy Appstore.xcscheme
@@ -56,6 +56,16 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
+      <BuildableProductRunnable
+         runnableDebuggingMode = "0">
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "E7B87E2225338121004358E3"
+            BuildableName = "Cloudy.app"
+            BlueprintName = "Cloudy"
+            ReferencedContainer = "container:Cloudy.xcodeproj">
+         </BuildableReference>
+      </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Appstore">
