+++ Cloudy/Model/LaunchArgumentHandler.swift
@@ -1,6 +1,7 @@
 // Copyright (c) 2020 Nomad5. All rights reserved.
 
 import Foundation
+import UIKit
 
 /// Handle the injected launch arguments
 class LaunchArgumentHandler {
