+++ Cloudy/Moonlight/View/AbsoluteTouchHandler.h
@@ -1,19 +1,23 @@
 //
-//  RelativeTouchHandler.h
+//  AbsoluteTouchHandler.h
 //  Moonlight
 //
 //  Created by Cameron Gutman on 11/1/20.
 //  Copyright © 2020 Moonlight Game Streaming Project. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #import "StreamView.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface RelativeTouchHandler : UIResponder
+@interface AbsoluteTouchHandler : UIResponder
 
     - (id)initWithView:(StreamView *)view;
 
 @end
 
 NS_ASSUME_NONNULL_END
+
+#endif
