+++ Cloudy/Moonlight/View/Cloudy-Bridging-Header.h
@@ -2,6 +2,9 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
+#include "InputPresenceDelegate.h"
+#include "UserInteractionDelegate.h"
+#include "OnScreenControlsLevel.h"
 #include "OnScreenControls.h"
 #include "ControllerSupport.h"
 #include "StreamView.h"
