+++ Cloudy/Moonlight/View/Controller.m
@@ -6,8 +6,12 @@
 //  Copyright © 2019 Moonlight Game Streaming Project. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #include "Controller.h"
 
 @implementation Controller
 
 @end
+
+#endif
