+++ Cloudy/Moonlight/View/HapticContext.m
@@ -6,6 +6,8 @@
 //  Copyright © 2020 Moonlight Game Streaming Project. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #import "HapticContext.h"
 #import "Log.h"
 
@@ -173,3 +175,5 @@
     }
 
 @end
+
+#endif
