+++ Cloudy/Moonlight/View/OnScreenControls.m
@@ -6,6 +6,8 @@
 //  Copyright (c) 2014 Moonlight Stream. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #import "OnScreenControls.h"
 #import "StreamView.h"
 #import "ControllerSupport.h"
@@ -1175,3 +1177,5 @@
     }
 
 @end
+
+#endif
