+++ Cloudy/Moonlight/View/RelativeTouchHandler.m
@@ -6,6 +6,8 @@
 //  Copyright © 2020 Moonlight Game Streaming Project. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #import "RelativeTouchHandler.h"
 #import "Log.h"
 
@@ -223,3 +225,5 @@ static const int REFERENCE_HEIGHT = 720;
 #endif
 
 @end
+
+#endif
