+++ Cloudy/Moonlight/View/StreamView.m
@@ -6,6 +6,8 @@
 //  Copyright (c) 2014 Moonlight Stream. All rights reserved.
 //
 
+#ifdef NON_APPSTORE
+
 #import "StreamView.h"
 #include "Limelight.h"
 #import "ControllerSupport.h"
@@ -767,3 +769,5 @@ static const double X1_MOUSE_SPEED_DIVISOR = 2.5;
 
 
 @end
+
+#endif
